순환 참조 제거 및 리펙토링
