+++ src/main/java/com/cozybinarybase/accountstopthestore/model/message/config/WebSocketConfig.java
@@ -24,12 +24,12 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
   @Override
   public void registerStompEndpoints(StompEndpointRegistry registry) {
     registry.addEndpoint("/ws")
-        .setAllowedOriginPatterns("*");
-//        .withSockJS();
+        .setAllowedOriginPatterns("*")
+        .withSockJS();
   }
 
   @Override
   public void configureClientInboundChannel(ChannelRegistration registration) {
-//    registration.interceptors(stompHandler);
+    registration.interceptors(stompHandler);
   }
 }
